Path Update and Fixes
Fixed path update so tutorial and challenge button and admin home page work for the user and admin homepage, tutorial, and challenge pages. NOTE: We will be using the user tutorial and challenge pages for the admin's access to the challenges and tutorials they made and challenges and tutorials that others have made.